Update LBM.hpp
improved variables allocation

+++ code/LBM.hpp
@@ -16,16 +16,16 @@
 class LBmethod {
 private:
     // Parameters
-    const unsigned int NSTEPS;
-    const unsigned int NX;
-    const unsigned int NY;
+    const size_t NSTEPS;
+    const size_t NX;
+    const size_t NY;
     const double u_lid;
     const double Re;
     double u_lid_dyn;
-    const unsigned int num_cores;
+    const size_t num_cores;
 
     // Fixed parameters
-    const unsigned int ndirections = 9;
+    const size_t ndirections = 9;
     const double tau;
     const double sigma = 10.0;
 
@@ -43,12 +43,12 @@ private:
     std::vector<double> f_temp;
 
     // Overloaded function for 2D to 1D indexing
-    inline int INDEX(int x, int y, int NX) {
+    inline size_t INDEX(size_t x, size_t y, size_t NX) {
         return x + NX * y;
     }
 
     // Overloaded function for 3D to 1D indexing
-    inline int INDEX(int x, int y, int i, int NX, int ndirections) {
+    inline size_t INDEX(size_t x, size_t y, size_t i, size_t NX, size_t ndirections) {
         return i + ndirections * (x + NX * y);
     }
 
@@ -56,7 +56,7 @@ private:
 
 public:
     // Constructor
-    LBmethod(const unsigned int NSTEPS, const unsigned int NX, const unsigned int NY, const double u_lid, const double Re, const unsigned int num_cores);
+    LBmethod(const size_t NSTEPS, const size_t NX, const size_t NY, const double u_lid, const double Re, const size_t num_cores);
 
     // Methods
     void Initialize();
@@ -65,7 +65,7 @@ public:
     void Collisions();
     void Streaming();
     void Run_simulation();
-    void Visualization(unsigned int t);
+    void Visualization(size_t t);
 };
 
 #endif // LBMETHOD_H
